requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart
reportlab
slowapi>=0.1.9
psutil>=5.9
aiofiles>=23.0
redis>=5.0.0
httpx>=0.27.0
uvloop>=0.19.0
openpyxl>=3.1.0
sentry-sdk[fastapi]>=2.0.0   # optional — only used when SENTRY_DSN is set
